+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
-## 2.0.1
+## 2.0.2
+- experimental [XARF](http://xarf.org/) reports reading
+## 2.0.1 (2022-11-21)
 - fix #24 stable locale
 - allow `file-magic` instead of `python-magic` package be installed on the system (#25)
 - fix GPG signing when attaching file with name longer than 34 chars (#19)